fix mo09/ex02

+++ mo09/ex02/benchmark.cpp
@@ -6,7 +6,7 @@
 /*   By: pepaloma <pepaloma@student.42urduliz.com>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/12 15:22:02 by pepaloma          #+#    #+#             */
-/*   Updated: 2025/05/15 17:20:06 by pepaloma         ###   ########.fr       */
+/*   Updated: 2025/05/20 12:32:43 by pepaloma         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,24 +17,24 @@
 #include <ctime>
 #include "PmergeMe.hpp"
 
-long benchmark_vector(std::vector<std::string> args, PmergeMe::Vector& result)
+double benchmark_vector(std::vector<std::string> args, PmergeMe::Vector& result)
 {
 	std::clock_t start = std::clock();
 	parse_vector(args, result);
 	if (result.size() > 1)
 		result.mergeInsertSort();
 	std::clock_t end = std::clock();
-	long time = static_cast<long>((static_cast<double>(end - start) / CLOCKS_PER_SEC) * 1e6);
+	double time = (static_cast<double>(end - start) / CLOCKS_PER_SEC) * 1e6;
 	return (time);
 }
 
-long benchmark_list(std::vector<std::string> args, PmergeMe::List& result)
+double benchmark_list(std::vector<std::string> args, PmergeMe::List& result)
 {
 	std::clock_t start = std::clock();
 	parse_list(args, result);
 	if (result.size() > 1)
 		result.mergeInsertSort();
 	std::clock_t end = std::clock();
-	long time = static_cast<long>((static_cast<double>(end - start) / CLOCKS_PER_SEC) * 1e6);
+	double time = (static_cast<double>(end - start) / CLOCKS_PER_SEC) * 1e6;
 	return (time);
 }
