+++ mo09/ex02/main.cpp
@@ -6,7 +6,7 @@
 /*   By: pepaloma <pepaloma@student.42urduliz.com>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/01 13:55:00 by pepaloma          #+#    #+#             */
-/*   Updated: 2025/05/15 16:11:11 by pepaloma         ###   ########.fr       */
+/*   Updated: 2025/05/20 12:27:48 by pepaloma         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -50,9 +50,9 @@ int main(int argc, char **argv)
 	PmergeMe::Vector v_unsorted; parse_vector(args, v_unsorted);
 	PmergeMe::List l_unsorted; parse_list(args, l_unsorted);
 	PmergeMe::List l_sorted;
-	long vector_time = benchmark_list(args, l_sorted);
+	double vector_time = benchmark_list(args, l_sorted);
 	PmergeMe::Vector v_sorted;
-	long list_time = benchmark_vector(args, v_sorted);
+	double list_time = benchmark_vector(args, v_sorted);
 	std::cout << "Before:	"; PmergeMe::printNumbers(v_unsorted); std::cout << std::endl;
 	std::cout << "Before:	"; PmergeMe::printNumbers(l_unsorted); std::cout << std::endl;
 	std::cout << "After:	"; PmergeMe::printNumbers(v_sorted); std::cout << std::endl;
